+++ src/RcsbFvWeb/RcsbFvBuilder/RcsbFvInstanceBuilder.ts
@@ -5,9 +5,11 @@ import {TagDelimiter} from "../../RcsbUtils/TagDelimiter";
 import {PolymerEntityInstanceInterface} from "../../RcsbCollectTools/Translators/PolymerEntityInstancesCollector";
 import {RcsbFvCoreBuilder} from "./RcsbFvCoreBuilder";
 import {rcsbFvCtxManager} from "./RcsbFvContextManager";
-import {GroupOptionPropsInterface, OptionPropsInterface, SelectOptionInterface} from "../WebTools/SelectButton";
+import {
+    SelectOptionInterface,
+    SelectOptionProps
+} from "../WebTools/SelectButton";
 import {RcsbFvModulePublicInterface} from "../RcsbFvModule/RcsbFvModuleInterface";
-import {OptionProps} from "react-select/dist/declarations/src/components/Option";
 import {RcsbFvInterface} from "../RcsbFvModule/RcsbFvInterface";
 
 export interface InstanceSequenceOnchangeInterface {
@@ -24,7 +26,7 @@ export interface InstanceSequenceConfig {
     onChangeCallback?:(x: InstanceSequenceOnchangeInterface)=>void;
     filterInstances?: Set<string>;
     displayAuthId?: boolean;
-    selectButtonOptionProps?: (props: OptionProps<OptionPropsInterface,false,GroupOptionPropsInterface>)=>JSX.Element;
+    selectButtonOptionProps?: (props: SelectOptionProps)=>JSX.Element;
     module?: instanceModule;
 }